# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brain Tumor Detection System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  error?: string;
}

export interface MessageResponse {
  message: string;
  success: boolean;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  admin: "admin",
  doctor: "doctor",
  student: "student",
} as const;

export interface RegisterRequest {
  fullName: string;
  email: string;
  password: string;
  confirmPassword: string;
  organization?: string;
  role?: RegisterRequestRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  doctor: "doctor",
  student: "student",
} as const;

export interface User {
  id: number;
  fullName: string;
  email: string;
  role: UserRole;
  organization?: string;
  avatar?: string;
  isActive: boolean;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface ForgotPasswordRequest {
  email: string;
}

export interface VerifyCodeRequest {
  email: string;
  code: string;
}

export interface ResetPasswordRequest {
  email: string;
  code: string;
  newPassword: string;
  confirmPassword: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
  confirmPassword: string;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  doctor: "doctor",
  student: "student",
} as const;

export interface UpdateUserRequest {
  fullName?: string;
  email?: string;
  organization?: string;
  role?: UpdateUserRequestRole;
  isActive?: boolean;
}

export interface UsersListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export type ScanStatus = (typeof ScanStatus)[keyof typeof ScanStatus];

export const ScanStatus = {
  pending: "pending",
  analyzed: "analyzed",
  reviewed: "reviewed",
} as const;

export interface Scan {
  id: number;
  imageUrl?: string;
  status: ScanStatus;
  patientId: number;
  patientName?: string;
  uploadedById: number;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface ScanNote {
  id: number;
  scanId: number;
  doctorId: number;
  doctorName?: string;
  note: string;
  createdAt: string;
}

export interface AddNoteRequest {
  note: string;
}

export type ScanWithAnalysisStatus =
  (typeof ScanWithAnalysisStatus)[keyof typeof ScanWithAnalysisStatus];

export const ScanWithAnalysisStatus = {
  pending: "pending",
  analyzed: "analyzed",
  reviewed: "reviewed",
} as const;

export type AnalysisResultTumorType =
  (typeof AnalysisResultTumorType)[keyof typeof AnalysisResultTumorType];

export const AnalysisResultTumorType = {
  glioma: "glioma",
  meningioma: "meningioma",
  pituitary: "pituitary",
  no_tumor: "no_tumor",
} as const;

export type AnalysisResultRiskLevel =
  (typeof AnalysisResultRiskLevel)[keyof typeof AnalysisResultRiskLevel];

export const AnalysisResultRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface AnalysisResult {
  id: number;
  scanId: number;
  tumorType: AnalysisResultTumorType;
  confidence: number;
  riskLevel: AnalysisResultRiskLevel;
  summary?: string;
  keyFindings?: string[];
  analyzedAt: string;
}

export interface ScanWithAnalysis {
  id: number;
  imageUrl?: string;
  status: ScanWithAnalysisStatus;
  patientId: number;
  patientName?: string;
  uploadedById: number;
  notes?: string;
  createdAt: string;
  updatedAt: string;
  analysis?: AnalysisResult;
  doctorNotes: ScanNote[];
}

export interface ScansListResponse {
  scans: Scan[];
  total: number;
  page: number;
  limit: number;
}

export interface AnalysisListResponse {
  analyses: AnalysisResult[];
  total: number;
  page: number;
  limit: number;
}

export type DashboardStatsTumorTypeDistributionItem = {
  type: string;
  count: number;
};

export interface DashboardStats {
  totalScans: number;
  totalUsers: number;
  totalDoctors: number;
  totalStudents: number;
  pendingScans: number;
  analyzedScans: number;
  malignantCases: number;
  benignCases: number;
  recentScans: Scan[];
  tumorTypeDistribution: DashboardStatsTumorTypeDistributionItem[];
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  scan_result: "scan_result",
  malfunction: "malfunction",
  ai_response: "ai_response",
  appointment: "appointment",
  app_update: "app_update",
} as const;

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationsListResponse {
  notifications: Notification[];
  unreadCount: number;
}

export type GetUsersParams = {
  role?: GetUsersRole;
  page?: number;
  limit?: number;
};

export type GetUsersRole = (typeof GetUsersRole)[keyof typeof GetUsersRole];

export const GetUsersRole = {
  admin: "admin",
  doctor: "doctor",
  student: "student",
} as const;

export type GetScansParams = {
  page?: number;
  limit?: number;
  status?: GetScansStatus;
  patientId?: number;
};

export type GetScansStatus =
  (typeof GetScansStatus)[keyof typeof GetScansStatus];

export const GetScansStatus = {
  pending: "pending",
  analyzed: "analyzed",
  reviewed: "reviewed",
} as const;

export type UploadScanBody = {
  file?: Blob;
  patientId?: number;
  notes?: string;
};

export type GetAllAnalysisParams = {
  page?: number;
  limit?: number;
};